requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
PyYAML>=5.4.0
h5py>=3.3.0
opencv-python>=4.5.0
pillow>=8.3.0
tensorboard>=2.7.0
wandb>=0.12.0
